Created Readme file.
There had been problems with the previous uploads of "Feather Color Quiz". Have tried to share my project as one Zip file on Git but looks like others was not able to see it. I deleted the entire repository on Git, but it looks like the name of the repository somehow stayed on my computer, and it refused to perform a push because it said the git.config wouldn't allow me. I resort to creating a new project with a different name, copied everything over from "Feather Color Quiz", except the app name of course.